requirements.txt
fastapi
uvicorn
pillow
requests
pydantic